flaskapp/requirements.txt
Flask~=1.1.2
PyJWT~=1.7.1
Flask-SQLAlchemy~=2.4.4
python-dotenv~=0.15.0
psycopg2-binary~=2.8.6
flask-sqlacodegen~=1.1.8
flask-marshmallow~=0.14.0
marshmallow-sqlalchemy~=0.24.2
flask_restful~=0.3.8
flask-jwt-extended